requirements.txt
flask
influxdb-client
python-dotenv